Completed days 1-5

+++ 2024/day_01/part2.py
@@ -4,7 +4,7 @@ list2 = []
 with open("input.txt", "r") as file:
     for line in file:
         num1, num2 = line.strip().split("   ")
-        num1, num2 = float(num1), float(num2)
+        num1, num2 = int(num1), int(num2)
 
         list1.append(num1)
         list2.append(num2)
